# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 드림코어 미로 타이쿤 API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface RegisterBody {
  username: string;
  password: string;
  nickname: string;
}

export interface LoginBody {
  username: string;
  password: string;
}

export interface User {
  id: number;
  username: string;
  nickname: string;
  role: string;
  currency: number;
  /** @nullable */
  equippedSkin?: string | null;
  isBanned: boolean;
  totalScore: number;
  playtime: number;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface UserWithPassword {
  id: number;
  username: string;
  nickname: string;
  password: string;
  role: string;
  currency: number;
  /** @nullable */
  equippedSkin?: string | null;
  isBanned: boolean;
  totalScore: number;
  playtime: number;
  createdAt: string;
}

export interface UpdateUserBody {
  nickname?: string;
  role?: string;
  currency?: number;
}

export interface BanUserBody {
  banned: boolean;
}

export interface InventoryItem {
  id: number;
  itemId: string;
  itemType: string;
  acquiredAt: string;
}

export type GameStateUpgrades = { [key: string]: unknown };

export type GameStatePosition = { [key: string]: unknown };

export type GameStateStats = { [key: string]: unknown };

export interface GameState {
  id: number;
  userId: number;
  level: number;
  currency: number;
  upgrades: GameStateUpgrades;
  position: GameStatePosition;
  stats: GameStateStats;
  updatedAt: string;
}

export type UpdateGameStateBodyUpgrades = { [key: string]: unknown };

export type UpdateGameStateBodyPosition = { [key: string]: unknown };

export type UpdateGameStateBodyStats = { [key: string]: unknown };

export interface UpdateGameStateBody {
  level?: number;
  currency?: number;
  upgrades?: UpdateGameStateBodyUpgrades;
  position?: UpdateGameStateBodyPosition;
  stats?: UpdateGameStateBodyStats;
}

export type UpgradeEffect = { [key: string]: unknown };

export interface Upgrade {
  id: string;
  name: string;
  description: string;
  cost: number;
  category: string;
  maxLevel: number;
  effect: UpgradeEffect;
}

export type SnapshotData = { [key: string]: unknown };

export interface Snapshot {
  id: number;
  userId: number;
  data: SnapshotData;
  createdAt: string;
}

export interface GameServer {
  id: number;
  name: string;
  ownerId: number;
  ownerNickname?: string;
  mode: string;
  maxPlayers: number;
  currentPlayers: number;
  isPublic: boolean;
  complexity: number;
  trapRate: number;
  mapWarp: boolean;
  status: string;
  createdAt: string;
}

export interface CreateServerBody {
  name: string;
  mode: string;
  maxPlayers?: number;
  isPublic?: boolean;
  complexity?: number;
  trapRate?: number;
  mapWarp?: boolean;
}

export interface UpdateServerBody {
  name?: string;
  maxPlayers?: number;
  isPublic?: boolean;
  complexity?: number;
  trapRate?: number;
  mapWarp?: boolean;
  status?: string;
}

export interface ServerStats {
  totalServers: number;
  activeServers: number;
  totalPlayers: number;
  onlinePlayers: number;
}

export interface ShopItem {
  id: string;
  name: string;
  description: string;
  type: string;
  price: number;
  rarity: string;
  color?: string;
  adminOnly: boolean;
}

export interface BuyItemBody {
  itemId: string;
}

export interface EquipSkinBody {
  /** @nullable */
  itemId: string | null;
}

export interface LeaderboardEntry {
  rank: number;
  userId: number;
  nickname: string;
  totalScore: number;
  playtime: number;
  /** @nullable */
  equippedSkin?: string | null;
}

export interface AdminStats {
  totalUsers: number;
  onlineUsers: number;
  totalServers: number;
  activeServers: number;
  totalCurrencyCirculating: number;
  totalReports: number;
  pendingReports: number;
  recentSignups: number;
}

export type LogEntryMetadata = { [key: string]: unknown };

export interface LogEntry {
  id: number;
  type: string;
  message: string;
  /** @nullable */
  userId?: number | null;
  metadata?: LogEntryMetadata;
  createdAt: string;
}

export interface AnnouncementBody {
  message: string;
  channel?: string;
}

export interface WorldSettings {
  complexity: number;
  wallThickness: number;
  trapRate: number;
  mapWarp: boolean;
  entitySpawnRate: number;
  ambientDarkness: number;
  glitchFrequency: number;
}

export interface AnomalyLog {
  id: number;
  userId: number;
  nickname: string;
  type: string;
  detail: string;
  severity: string;
  createdAt: string;
}

export interface Report {
  id: number;
  reporterId: number;
  reporterNickname?: string;
  targetNickname: string;
  reason: string;
  status: string;
  createdAt: string;
}

export interface CreateReportBody {
  targetNickname: string;
  reason: string;
}

export interface UpdateReportBody {
  status: string;
}

export interface ChatMessage {
  id: number;
  userId: number;
  nickname: string;
  channel: string;
  message: string;
  createdAt: string;
}

export interface GuestbookMessage {
  id: number;
  userId: number;
  nickname: string;
  message: string;
  wallId: string;
  createdAt: string;
}

export interface CreateGuestbookBody {
  message: string;
  wallId: string;
}

export type GetAdminLogsParams = {
  limit?: number;
  type?: string;
};

export type GetChatLogsParams = {
  channel?: string;
  limit?: number;
};