# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title
  title: Api
  version: 0.1.0
  description: 드림코어 미로 타이쿤 API
servers:
  - url: /api
    description: API 기본 경로
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: 회원가입
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: 가입 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: 잘못된 요청
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: 로그인
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: 로그인 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: 인증 실패
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: 현재 로그인한 유저 정보
      responses:
        "200":
          description: 유저 정보
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: 미인증
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ───────────────── USERS ─────────────────
...
  /game/state:
    get:
      operationId: getGameState
      tags: [game]
      summary: 내 게임 상태 조회
      responses:
        "200":
          description: 게임 상태
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameState"
    patch:
      operationId: updateGameState
      tags: [game]
      summary: 게임 상태 저장
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGameStateBody"
      responses:
        "200":
          description: 저장된 상태
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameState"
...
  /game/upgrades/{upgradeId}/purchase:
    post:
      operationId: purchaseUpgrade
      tags: [game]
      summary: 업그레이드 구매
      parameters:
        - name: upgradeId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 구매 완료
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameState"
        "400":
          description: 잔액 부족 또는 이미 구매함
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /game/snapshot/{snapshotId}/rollback:
    post:
      operationId: rollbackSnapshot
      tags: [game]
      summary: 스냅샷으로 롤백
      parameters:
        - name: snapshotId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: 롤백 완료
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ───────────────── SERVERS ─────────────────
...
  /admin/logs:
    get:
      operationId: getAdminLogs
      tags: [admin]
      summary: 서버 로그 목록
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
        - name: type
          in: query
          schema:
            type: string
      responses:
        "200":
          description: 로그 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LogEntry"
  /admin/announce:
    post:
      operationId: sendAnnouncement
      tags: [admin]
      summary: 전체 공지 전송
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnouncementBody"
      responses:
        "200":
          description: 전송 완료
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/world-settings:
    get:
      operationId: getWorldSettings
      tags: [admin]
      summary: 월드 설정 조회
      responses:
        "200":
          description: 월드 설정
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorldSettings"
    patch:
      operationId: updateWorldSettings
      tags: [admin]
      summary: 월드 설정 변경
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorldSettings"
      responses:
        "200":
          description: 변경된 설정
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorldSettings"
...
  /chat/logs:
    get:
      operationId: getChatLogs
      tags: [chat]
      summary: 채팅 로그 조회
      parameters:
        - name: channel
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: 채팅 로그
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"

  # ───────────────── GUESTBOOK ─────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # ── Auth Schemas ──
    RegisterBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        nickname:
          type: string
      required: [username, password, nickname]
    LoginBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required: [username, password]
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required: [token, user]

    # ── User Schemas ──
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        nickname:
          type: string
        role:
          type: string
        currency:
          type: integer
        equippedSkin:
          type: ["string", "null"]
        isBanned:
          type: boolean
        totalScore:
          type: integer
        playtime:
          type: integer
        createdAt:
          type: string
      required: [id, username, nickname, role, currency, isBanned, totalScore, playtime, createdAt]
...
    GameState:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        level:
          type: integer
        currency:
          type: integer
        upgrades:
          type: object
        position:
          type: object
        stats:
          type: object
        updatedAt:
          type: string
      required: [id, userId, level, currency, upgrades, position, stats, updatedAt]
    UpdateGameStateBody:
      type: object
      properties:
        level:
          type: integer
        currency:
          type: integer
        upgrades:
          type: object
        position:
          type: object
        stats:
          type: object
...
    LogEntry:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
        message:
          type: string
        userId:
          type: ["integer", "null"]
        metadata:
          type: object
        createdAt:
          type: string
      required: [id, type, message, createdAt]
    AnnouncementBody:
      type: object
      properties:
        message:
          type: string
        channel:
          type: string
      required: [message]
    WorldSettings:
      type: object
      properties:
        complexity:
          type: integer
        wallThickness:
          type: number
        trapRate:
          type: number
        mapWarp:
          type: boolean
        entitySpawnRate:
          type: number
        ambientDarkness:
          type: number
        glitchFrequency:
          type: number
      required: [complexity, wallThickness, trapRate, mapWarp, entitySpawnRate, ambientDarkness, glitchFrequency]
...
    ChatMessage:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        nickname:
          type: string
        channel:
          type: string
        message:
          type: string
        createdAt:
          type: string
      required: [id, userId, nickname, channel, message, createdAt]

    # ── Guestbook ──
...